Added tests for NewTabComponent and PopUpComponent
