+++ package.json
@@ -2,7 +2,7 @@
 	"name": "mindful",
 	"version": "1.0.0",
 	"description": "Store lists of bookmarks in separate categories",
-	"main": "Index.js",
+	"main": "NewTab.js",
 	"scripts": {
 		"dev": "vite",
 		"build": "vite build",
@@ -59,8 +59,8 @@
 		"esbuild": "^0.25.6",
 		"html-webpack-plugin": "^5.4.0",
 		"identity-obj-proxy": "^3.0.0",
-		"jest": "^30.0.4",
-		"jest-environment-jsdom": "^30.0.4",
+		"jest": "^30.0.5",
+		"jest-environment-jsdom": "^30.0.5",
 		"style-loader": "^3.3.2",
 		"terser-webpack-plugin": "^5.3.11",
 		"tsx": "^4.20.3",
