+++ public/NewTab.html
@@ -3,11 +3,11 @@
   <head>
     <meta charset="UTF-8" />
     <title>Mindful</title>
-    <link rel="stylesheet" href="index.css" />
+    <link rel="stylesheet" href="newtab.css" />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" integrity="sha512-Fo3rlrZj/k7ujTnHg4CGR2D7kSs0v4LLanw2qksYuRlEzO+tcaEPQogQ0KaoGN26/zrn20ImR1DfuLWnOo7aBA==" crossorigin="anonymous" referrerpolicy="no-referrer" />
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="./index.js"></script>
+    <script type="module" src="./newtab.js"></script>
   </body>
 </html>
