+++ public/manifest.json
@@ -5,7 +5,7 @@
   "version": "1.0.0",
 
   "chrome_url_overrides": {
-    "newtab": "Index.html"
+    "newtab": "NewTab.html"
   },
 
   "icons": {
