+++ src/styles/AddBookmarkInline.css
@@ -1,4 +1,4 @@
-@import "Index.css";
+@import "NewTab.css";
 
 .add-link-inline-container {
   position: "relative";
