+++ src/styles/EditableBookmarkGroupHeading.css
@@ -1,4 +1,4 @@
-@import "Index.css";
+@import "NewTab.css";
 
 .editable-heading {
   font-size: 24px;
