+++ vite.config.js
@@ -8,7 +8,7 @@ export default defineConfig({
     build: {
         rollupOptions: {
             input: {
-                index: './src/index.jsx', // Entry point for the new tab page
+                newtab: './src/newtab.jsx', // Entry point for the new tab page
                 popup: './src/popup.jsx',   // Entry point for the popup window
             },
             output: {
